Trying to fix cirrus(3)
