+++ app/src/androidTest/java/io/github/polysmee/zroomActivityTests/RoomActivityTest.java
@@ -1,10 +1,10 @@
-package io.github.polysmee.roomActivityTests;
+package io.github.polysmee.zroomActivityTests;
 
-import android.app.Activity;
 import android.content.Intent;
 
 import androidx.test.core.app.ActivityScenario;
 import androidx.test.core.app.ApplicationProvider;
+import androidx.test.espresso.intent.Intents;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import com.google.android.gms.tasks.Tasks;
@@ -17,25 +17,29 @@ import org.junit.runner.RunWith;
 import java.util.concurrent.TimeUnit;
 
 import io.github.polysmee.R;
+import io.github.polysmee.appointments.AppointmentActivity;
 import io.github.polysmee.database.DatabaseFactory;
 import io.github.polysmee.login.AuthenticationFactory;
 import io.github.polysmee.login.MainUser;
 import io.github.polysmee.room.RoomActivity;
 import io.github.polysmee.znotification.AppointmentReminderNotification;
 
+import static androidx.test.espresso.intent.Intents.intended;
+import static androidx.test.espresso.intent.matcher.IntentMatchers.hasExtra;
+import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertContains;
 import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
-import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
+import static com.schibsted.spain.barista.interaction.BaristaMenuClickInteractions.clickMenu;
 import static com.schibsted.spain.barista.interaction.BaristaSleepInteractions.sleep;
-import static org.junit.Assert.assertEquals;
+import static com.schibsted.spain.barista.interaction.BaristaViewPagerInteractions.swipeViewPagerForward;
 
 @RunWith(AndroidJUnit4.class)
-public class RoomActivityNotParticipantTest {
+public class RoomActivityTest {
     private static final String username1 = "Mathis L'utilisateur";
-    private static final String id2 = "posdkojerzyugcwxu";
+    private static final String id2 = "bxcwviusergpoza";
     private static final String username2 = "Sami L'imposteur";
 
     private static final String appointmentTitle = "It's a title";
-    private static final String appointmentId = "oiuowfpkksdnf";
+    private static final String appointmentId = "cwxbihezroijgdf";
     private static final String appointmentCourse = "Totally not SWENG";
     private static final long appointmentStart = 265655445;
 
@@ -47,39 +51,56 @@ public class RoomActivityNotParticipantTest {
         AuthenticationFactory.setTest();
         FirebaseApp.clearInstancesForTest();
         FirebaseApp.initializeApp(ApplicationProvider.getApplicationContext());
-        Tasks.await(AuthenticationFactory.getAdaptedInstance().createUserWithEmailAndPassword("RoomActivityNotParticipantTest@gmail.com", "fakePassword"));
+        Tasks.await(AuthenticationFactory.getAdaptedInstance().createUserWithEmailAndPassword("RoomActivityTest@gmail.com", "fakePassword"));
         DatabaseFactory.getAdaptedInstance().getReference("users").child(MainUser.getMainUser().getId()).child("name").setValue(username1);
         DatabaseFactory.getAdaptedInstance().getReference("users").child(id2).child("name").setValue(username2);
 
         DatabaseFactory.getAdaptedInstance().getReference("appointments").child(appointmentId).child("title").setValue(appointmentTitle);
         DatabaseFactory.getAdaptedInstance().getReference("appointments").child(appointmentId).child("course").setValue(appointmentCourse);
         DatabaseFactory.getAdaptedInstance().getReference("appointments").child(appointmentId).child("start").setValue(appointmentStart);
+        DatabaseFactory.getAdaptedInstance().getReference("appointments").child(appointmentId).child("participants").child(MainUser.getMainUser().getId()).setValue(true);
         DatabaseFactory.getAdaptedInstance().getReference("appointments").child(appointmentId).child("participants").child(id2).setValue(true);
     }
 
 
     @Test
-    public void onlyParticipantCanJoinARoom() {
+    public void titleOfTheActivityShouldBeTheAppointmentTitle() {
         Intent intent = new Intent(ApplicationProvider.getApplicationContext(), RoomActivity.class);
         intent.putExtra(RoomActivity.APPOINTMENT_KEY, appointmentId);
 
         try (ActivityScenario<RoomActivity> ignored = ActivityScenario.launch(intent)) {
-            sleep(2, TimeUnit.SECONDS);
-            assertDisplayed(R.id.roomActivityRemovedDialogText);
-            assertDisplayed(R.id.roomActivityRemovedDialogQuitButton);
+            assertContains(appointmentTitle);
         }
     }
 
     @Test
-    public void quitButtonShouldFinishTheActivity() throws InterruptedException {
+    public void infoItemMenuShouldFireAnIntentWithTheCurrentAppointment() {
         Intent intent = new Intent(ApplicationProvider.getApplicationContext(), RoomActivity.class);
+
         intent.putExtra(RoomActivity.APPOINTMENT_KEY, appointmentId);
 
-        ActivityScenario<RoomActivity> scenario = ActivityScenario.launch(intent);
-        sleep(1, TimeUnit.SECONDS);
-        clickOn(R.id.roomActivityRemovedDialogQuitButton);
-        assertEquals(Activity.RESULT_CANCELED, scenario.getResult().getResultCode());
-        Thread.sleep(1000);
-        scenario.close();
+        try (ActivityScenario<RoomActivity> ignored = ActivityScenario.launch(intent)) {
+            Intents.init();
+            clickMenu(R.id.roomMenuInfo);
+            intended(hasExtra(AppointmentActivity.LAUNCH_MODE, AppointmentActivity.DETAIL_MODE));
+            intended(hasExtra(AppointmentActivity.APPOINTMENT_ID, appointmentId));
+            Intents.release();
+        }
+    }
+
+    @Test
+    public void participantsAreCorrectlyDisplayed() {
+        Intent intent = new Intent(ApplicationProvider.getApplicationContext(), RoomActivity.class);
+
+        intent.putExtra(RoomActivity.APPOINTMENT_KEY, appointmentId);
+
+        try (ActivityScenario<RoomActivity> ignored = ActivityScenario.launch(intent)) {
+            swipeViewPagerForward();
+            sleep(1, TimeUnit.SECONDS);
+            swipeViewPagerForward();
+            sleep(2, TimeUnit.SECONDS);
+            assertDisplayed("You");
+            assertDisplayed(username2);
+        }
     }
 }
