+++ app/src/androidTest/java/io/github/polysmee/zroomActivityTests/RoomActivityVideoFragmentTest.java
@@ -1,4 +1,4 @@
-package io.github.polysmee.roomActivityTests;
+package io.github.polysmee.zroomActivityTests;
 
 
 import android.content.Intent;
