+++ app/src/androidTest/java/io/github/polysmee/zroomActivityTests/TestMessages.java
@@ -1,4 +1,4 @@
-package io.github.polysmee.roomActivityTests;
+package io.github.polysmee.zroomActivityTests;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
