+++ app/src/main/java/io/github/polysmee/profile/UserItemAutocomplete.java
@@ -1,9 +1,12 @@
 package io.github.polysmee.profile;
 
 public class UserItemAutocomplete {
-    private final String username;
-    private final String pictureId;
+    private String username;
+    private String pictureId;
 
+    public UserItemAutocomplete(){
+        //empty constructor
+    }
     public UserItemAutocomplete(String username,String pictureId){
         this.username = username;
         this.pictureId = pictureId;
@@ -14,5 +17,12 @@ public class UserItemAutocomplete {
     }
     public String getPictureId(){
         return pictureId;
+    }
+
+    public void setPictureId(String pictureId) {
+        this.pictureId = pictureId;
+    }
+    public  void setUsername(String username){
+        this.username = username;
     }
 }
